requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-stubs==1.13.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
idna==3.4
mypy==0.991
mypy-extensions==0.4.3
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.12
